Refactor Danger so that Hazards may use DangerTemplate paradigm.
Enemy still works the old way; can refactor in time if merited.  This way is improved because we store the data as fields and not as literals in overridden methods, as I discovered I had been doing.
Hazard shows how to create a template for creating Dangers without using an Enum.  We can't reference enemies Types, but we weren't doing that anyways, and doing so would probably violate encapsulation.

Notice NameAlgorithm is used in place of Lambdas, as we aren't using Java 1.8 (Android doesn't yet support).

Commented out code has been left in Hazard, will clean up before commit.